[clean] Integration LoginBundle to AppBundle
