+++ app/config/routing.yml
@@ -1,8 +1,3 @@
-o_auth_login:
-    resource: "@LoginBundle/Controller/"
-    type:     annotation
-    prefix:   /
-
 app:
     resource: "@AppBundle/Controller/"
     type:     annotation
