--- src/LoginBundle/LoginBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace LoginBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class LoginBundle extends Bundle
-{
-}
