--- src/LoginBundle/Tests/Controller/DefaultControllerTest.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace LoginBundle\Tests\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class DefaultControllerTest extends WebTestCase
-{
-    public function testLoginPageRender()
-    {
-        $client = static::createClient();
-
-        $crawler = $client->request('GET', '/login');
-
-        $this->assertEquals(200, $client->getResponse()->getStatusCode());
-
-        $this->assertTrue($crawler->filter('body.login')->count() > 0);
-    }
-}
